Use double instead of float
For better consistency use double everywhere instead of float. Otherwise it is very difficult to keep track where float or double has to be used (when dealing with ROOT histograms is not always apparent which one needs to be used). It may be interesting to evaluate the performance loss.

+++ include/wgDetectorDimension.h
@@ -43,8 +43,8 @@ public:
   int   GetPlnMax(int mod, int view, int pln, int axis);
   bool  GetPos_TwoDimRecon(int mod, int view, int reconpln, int reconch, int axis,
 	   		                                double* posxy, double* posz);
-  float GetScintiWidth(int mod, int view, int reconpln, int reconch, int axis);
-  float GetScintiThick(int mod, int view, int reconpln, int reconch, int axis);
+  double GetScintiWidth(int mod, int view, int reconpln, int reconch, int axis);
+  double GetScintiThick(int mod, int view, int reconpln, int reconch, int axis);
 
 };
 #endif
